+++ src/node/static-stubs/react-component.js
@@ -1,0 +1,3 @@
+module.exports = function StubbedComponent() {
+  return null;
+};